requirements.txt
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.3
Django==3.1.1
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-model-utils==4.0.0
django-notifications-hq==1.6.0
isort==5.5.2
jsonfield==3.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
